AHK: Add screen time functionality
This is an experimental addition to view the daily screen time. It is
not very well tested yet.

Time will be saved periodically to an .ini file or on reload. It will
stop recording time if there was an idle time of 30 seconds. The total
daily time is shown in the bottom left corner.